move to vite plus / oxc
