--- .prettierignore
@@ -1,4 +1,0 @@
-node_modules
-extension
-dist
-pnpm-lock.yaml
