--- .prettierrc
@@ -1,20 +1,0 @@
-{
-  "tabWidth": 2,
-  "useTabs": false,
-  "trailingComma": "all",
-  "printWidth": 100,
-  "singleQuote": true,
-  "quoteProps": "consistent",
-  "jsxSingleQuote": false,
-  "semi": true,
-  "arrowParens": "always",
-  "plugins": ["prettier-plugin-package"],
-  "overrides": [
-    {
-      "files": "sdk-components/**/prettierrc.js",
-      "options": {
-        "parser": "babel"
-      }
-    }
-  ]
-}
