+++ .vite-hooks/pre-commit
@@ -1,0 +1,1 @@
+vp staged
