+++ package.json
@@ -5,19 +5,19 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "build": "tsdown",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "prettier": "prettier . -c",
-    "prettier:fix": "prettier . -w",
-    "tsc": "tsc -p tsconfig.app.json --noEmit"
+    "build": "vp pack",
+    "lint": "vp lint .",
+    "lint:fix": "vp lint . --fix",
+    "format": "vp fmt . --check",
+    "format:fix": "vp fmt .",
+    "tsc": "tsc -p tsconfig.app.json --noEmit",
+    "prepare": "vp config"
   },
   "dependencies": {
     "@solid-primitives/scheduled": "1.5.3",
-    "@tanstack/solid-virtual": "^3.13.24",
+    "@tanstack/solid-virtual": "^3.13.25",
     "clsx": "2.1.1",
-    "eslint-plugin-simple-import-sort": "13.0.0",
-    "oniguruma-to-es": "4.3.4",
+    "oniguruma-to-es": "4.3.6",
     "solid-js": "1.9.13",
     "tailwindcss": "4.3.0",
     "vscode-textmate": "9.3.2",
@@ -25,28 +25,23 @@
   },
   "devDependencies": {
     "@changesets/cli": "2.31.0",
-    "@eslint/js": "9.39.2",
     "@microsoft/api-extractor": "7.52.11",
     "@tailwindcss/postcss": "4.3.0",
     "@tsdown/css": "0.22.0",
     "@types/archiver": "7.0.0",
     "@types/chrome": "0.1.42",
-    "@types/node": "25.8.0",
-    "@typescript-eslint/eslint-plugin": "8.59.3",
-    "@typescript-eslint/parser": "8.59.3",
+    "@types/node": "24.12.4",
     "archiver": "8.0.0",
-    "eslint": "9.33.0",
-    "eslint-config-prettier": "10.1.8",
     "eslint-plugin-solid": "0.14.5",
     "globals": "17.6.0",
-    "postcss": "8.5.14",
-    "prettier": "3.8.3",
-    "prettier-plugin-package": "2.0.0",
+    "oxfmt": "^0.51.0",
+    "oxlint": "^1.66.0",
+    "postcss": "8.5.15",
     "rollup-plugin-postcss": "4.0.2",
     "ts-node": "10.9.2",
-    "tsdown": "0.22.0",
     "typescript": "6.0.3",
-    "typescript-eslint": "8.59.3",
-    "vite-plugin-solid": "2.11.12"
-  }
+    "vite-plugin-solid": "2.11.12",
+    "vite-plus": "catalog:"
+  },
+  "packageManager": "pnpm@11.2.2"
 }
