--- todo.md
@@ -1,8 +1,0 @@
-view previous game-states
-to lock variables
-to view the current passage
-to view passage map
-to visit a passage
-to search variables/values
-notification when any variable changes to NaN
-speed hack
